Partial Completion
Worked up to (and a bit past) the deadline but still couldn't quite figure everything out. Right now the ball rotates just fine along one axis, but after it's rotated along that axis it flies out of the world if you try to rotate it along a different axis, even if it isn't rotating along that original axis anymore. I also might be getting the direction of rotation wrong? I'll look at that once I get the motion right.

My main hypothesis for the source of the issue is that something about the way the ball is rotating means that translations aren't functioning as expected; they might not be translating along that clean coordinate system anymore.

I'm planning to stop by office hours on Monday to get help about this.